+++ amazonorders/constants.py
@@ -53,3 +53,9 @@ class Constants:
         "User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) "
                       "Chrome/120.0.0.0 Safari/537.36",
     }
+
+    ##########################################################################
+    # Formats
+    ##########################################################################
+
+    VALID_DATE_FORMATS = ["%b %d, %Y", "%B %d, %Y"]